# swagger_local.json
{
  "swagger": "2.0",
  "info": {
    "description": "Ecommerce Api server for admin and web",
    "version": "1.0.0",
    "title": "Ecommerce"
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/admin/auth/login/admin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login to admin system",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Admin username and password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/login/user": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Member login check",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "member username and password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/member/admin": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Get all admins",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/member/admin/{adminId}": {
      "get": {
        "tags": [
          "admin member"
        ],
        "summary": "Find admin by id",
        "description": "Returns a single admin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "put": {
        "tags": [
          "admin member"
        ],
        "summary": "update admin by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterAdmin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "admin member"
        ],
        "summary": "Deletes an admin",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "adminId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
    "/admin/member/user": {
      "get": {
        "tags": [
          "admin user"
        ],
        "summary": "admin",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/member/user/{userId}": {
      "get": {
        "tags": [
          "admin user"
        ],
        "summary": "admin",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of admin to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "admin user"
        ],
        "summary": "update user by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "member update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "422": {
            "description": "Invalid ID supplied"
          }
        }
      },
      "delete": {
        "tags": [
          "admin user"
        ],
        "summary": "Deletes an admin",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/api/v1/verify": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Member verify",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/seller": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "get seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "post": {
        "tags": [
          "seller"
        ],
        "summary": "post seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "add seller",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/admin/seller/{sellerId}": {
      "put": {
        "tags": [
          "seller"
        ],
        "summary": "update seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sellerId",
            "in": "path",
            "description": "ID of seller to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "update seller",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "seller"
        ],
        "summary": "delete seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sellerId",
            "in": "path",
            "description": "ID of seller to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/seller": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "get seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      },
      "post": {
        "tags": [
          "seller"
        ],
        "summary": "post seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "add seller",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
    "/api/v1/seller/{sellerId}": {
      "put": {
        "tags": [
          "seller"
        ],
        "summary": "update seller",
        "description": "",
        "operationId": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "sellerId",
            "in": "path",
            "description": "ID of seller to update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "update seller",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Seller"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Invalid"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Seller": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "jeetwin"
        },
        "sellerDescription": {
          "type": "string",
          "example": "Game Points dfkjkdufb dakjbdhfkasf asfjuhaskfasbf afjkjafhg"
        },
        "accountName": {
          "type": "string",
          "example": "ram chanfra"
        },
        "accountNumber": {
          "type": "string",
          "example": "874259635826"
        }
      }
    },
...
    "UpdateUser": {
      "type": "object",
      "properties": {
        "fname": {
          "type": "string",
          "example": "himal"
        },
        "lname": {
          "type": "string",
          "example": "shrestha"
        },
        "city": {
          "type": "string",
          "example": "circle"
        },
        "state": {
          "type": "string",
          "example": "somwhere wlse"
        },
        "zip": {
          "type": "string",
          "example": "8945"
        },
        "phone": {
          "type": "string",
          "example": "1223234234"
        },
        "country": {
          "type": "string",
          "example": "nepal"
        },
        "address": {
          "type": "string",
          "example": "far"
        },
        "address2": {
          "type": "string",
          "example": "somewhere"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        },
        "role": {
          "type": "integer",
          "format": "int64",
          "example": "1",
          "description": "0 or 1"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "UpdateUser"
      }
    },
    "RegisterAdmin": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himalhs1@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "himal123"
        },
        "email": {
          "type": "string",
          "example": "himal123"
        },
        "roles": {
          "type": "string",
          "example": "himal123"
        },
        "fname": {
          "type": "string",
          "example": "himal123"
        },
        "lname": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "RegisterAdmin"
      }
    },
...
    "Auth": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string",
          "example": "himal"
        },
        "password": {
          "type": "string",
          "example": "himal"
        }
      },
      "xml": {
        "name": "Pet"
      }
    }
  },
...
}